feat(api): Add POST /api/articles/:article_id/comments endpoint
This endpoint enable a post request to add a comment to an article.
Request body accepts an object with 'username' and 'body' properties.
Responds with the posted comment.
Handle 400, and 500 errors and add corresponding tests.
Update the /api endpoint JSON file with the description of POST /api/articles/:article_id/comments endpoint.